Simplify Dependency API
